backend/requirements.txt
Flask
Pillow
pytesseract
openai
requests
jsonify
request